Put the BasicPasswordManagementTest.php in separate file. Suggest using regex for finding doubled words.
